fix(ui,api,db): address review findings for PRs 93-99 (#100) (#101)
Backend:
- Replace backref with back_populates on CreditReport/PrequalificationDecision (W-3)
- Add prequalification_superseded audit event before upsert delete (W-5)
- Change bankruptcy_flag to always False for mock data (W-6)
- Simplify monthly_debts None check in borrower_tools (W-11)
- Add no-financials guard to lo_issue_prequalification (W-29)
- Make DTI deterministic using per-product tracking dict (S-1)
- Move DISCLOSURE_BY_ID to module-level import (S-4)
- Add return_app param to get_application_status to avoid double fetch (S-5)
- Add asyncio.Lock for JWKS cache with double-check pattern (S-6)
- Remove unused sqlalchemy delete import

Frontend:
- Extract ChatBubble/TypingIndicator to shared chat-bubble component (W-22)
- Extract DOC_TYPE_LABELS/STAGE_BADGE to shared lib/labels.ts (W-24)
- Use ConditionSeveritySchema/ConditionStatusSchema enums in conditions schema (W-21)
- Add _streaming field to ChatMessage interface, remove type cast (W-27)
- Add role-based route guards with isRouteAllowed and ROLE_HOME redirect (W-15/W-18)
- Add keyboard accessibility to PipelineRow (tabIndex, role, onKeyDown, focus ring) (S-11)

Tests:
- Add app-not-found tests for all 3 prequal tools (W-28)
- Add no-financials rejection test (W-29)
- Add bankruptcy_flag always-false test (W-30)
- Add credit-too-low filtered product test (W-31)
- Add DTI determinism test (S-1)
- Add explicit prequalification_decision=None in data_factory (S-13)

+++ packages/api/src/schemas/application.py
@@ -16,7 +16,7 @@ class PrequalificationSummary(BaseModel):
 
     product_id: str
     product_name: str
-    max_loan_amount: Decimal
+    max_loan_amount: float
     estimated_rate: float
     issued_at: datetime
     expires_at: datetime
